requirements.txt
streamlit
gspread
google-auth
google-cloud-vision
google-api-python-client